requirements.txt
pandas
numpy
requests
xlsxwriter
pytz
supabase
python-dotenv
